Merge new multi-tenant code into master

+++ src/controllers/payments/admin/ExtraDelete.php
@@ -1,9 +1,14 @@
 <?php
 
 $db = app()->db;
+$tenant = app()->tenant;
+
 try {
-  $delete = $db->prepare("DELETE FROM extras WHERE ExtraID = ?");
-  $delete->execute([$id]);
+  $delete = $db->prepare("DELETE FROM extras WHERE ExtraID = ? AND Tenant = ?");
+  $delete->execute([
+    $id,
+    $tenant->getId()
+  ]);
   header("Location: " . autoUrl("payments/extrafees"));
 } catch (Exception $e) {
   halt(500);
